+++ pkg/install/init/models/input/model.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package domain
+package input
 
 import (
 	"fmt"
